Refactor Explorer into many components
fix issues from #2150 & use dark theme

Test Plan: existing test suite, manual testing
